Implemented the user handler

+++ internal/user/entity/user.go
@@ -3,7 +3,7 @@ package entity
 import "time"
 
 type User struct {
-	ID          int       `gorm:"primaryKey;autoIncrement;not null"`
+	ID          int64     `gorm:"primaryKey;autoIncrement;not null"`
 	Name        string    `gorm:"size:255;not null"`
 	Email       string    `gorm:"not null;unique"`
 	Batch       int       `gorm:"not null" `
